add sphinx extension to run examples
